Define real deps in lambda
