+++ src/functions/asyncCredential/TokenService/tokenService.test.ts
@@ -1,7 +1,6 @@
 import { KMSClient, VerifyCommand } from "@aws-sdk/client-kms";
 import { mockClient } from "aws-sdk-client-mock";
-import format from "ecdsa-sig-formatter";
-import { LogOrValue, log, value } from "../../types/logOrValue";
+import { TokenService } from "./tokenService";
 
 // Generated using jwt.io with their public and private keys
 const mockJwt =
@@ -52,37 +51,3 @@ describe("Token Service", () => {
     });
   });
 });
-
-export interface IValidateTokenPayload {
-  validateTokenPayload: (keyId: string, jwt: string) => LogOrValue<null>;
-}
-
-export interface IVerifyTokenSignature {
-  verifyTokenSignature: (
-    keyId: string,
-    jwt: string,
-  ) => Promise<LogOrValue<null>>;
-}
-
-export class TokenService implements IVerifyTokenSignature {
-  async verifyTokenSignature(
-    keyId: string,
-    jwt: string,
-  ): Promise<LogOrValue<null>> {
-    const [header, payload, signature] = jwt.split(".");
-    const kmsClient = new KMSClient();
-    const verifyCommand = new VerifyCommand({
-      KeyId: keyId,
-      SigningAlgorithm: "ECDSA_SHA_256",
-      Signature: format.joseToDer(signature, "ES256"),
-      Message: Buffer.from(`${header}.${payload}`),
-    });
-
-    const result = await kmsClient.send(verifyCommand);
-    if (result.SignatureValid !== true) {
-      return log("TOKEN_SIGNATURE_INVALID");
-    }
-
-    return value(null);
-  }
-}
